requirements.txt
django==5.1.3
djangorestframework==3.15.2
django-cors-headers==4.6.0
djangorestframework-simplejwt==5.3.1
django-filter==24.3
psycopg2-binary==2.9.10
python-dotenv==1.0.1
tensorflow==2.15.0
numpy==1.26.0
pandas==2.1.1
scikit-learn==1.2.2
Faker==20.0.0
